% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{print.bptr_bootstrap}
\alias{print.bptr_bootstrap}
\title{Print Method for bptr_bootstrap Objects}
\usage{
\method{print}{bptr_bootstrap}(x, ...)
}
\arguments{
\item{x}{A "bptr_bootstrap" object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for bptr_bootstrap Objects
}
