% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildIndicators}
\alias{buildIndicators}
\title{Build Threshold Indicators}
\usage{
buildIndicators(q, gamma)
}
\arguments{
\item{q}{Numeric vector of threshold variable}

\item{gamma}{Threshold value}
}
\value{
Numeric vector of indicators (1 if q <= gamma, 0 otherwise)
}
\description{
Build Threshold Indicators
}
