% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{computeSSR}
\alias{computeSSR}
\title{Compute SSR for 2-Regime Grid Search}
\usage{
computeSSR(y_dm, X_dm, g_vec, q_dm, buffer = FALSE)
}
\arguments{
\item{y_dm}{Demeaned dependent variable}

\item{X_dm}{Demeaned explanatory variables}

\item{g_vec}{Scalar or two-element vector of threshold candidate(s)}

\item{q_dm}{Demeaned threshold variable}

\item{buffer}{Logical}
}
\value{
Numeric SSR (Inf if any regime is too small)
}
\description{
Compute SSR for 2-Regime Grid Search
}
