% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concentratedOLS}
\alias{concentratedOLS}
\title{Concentrated OLS for 2-Regime Threshold Regression}
\usage{
concentratedOLS(y_dm, X_dm, ind1, ind2)
}
\arguments{
\item{y_dm}{Demeaned dependent variable}

\item{X_dm}{Demeaned explanatory variables matrix}

\item{ind1}{Indicator vector for regime 1}

\item{ind2}{Indicator vector for regime 2}
}
\value{
List with beta1, beta2, fitted, resid, sigma2
}
\description{
Concentrated OLS for 2-Regime Threshold Regression
}
