% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concentratedOLS3}
\alias{concentratedOLS3}
\title{Concentrated OLS for 3-Regime Threshold Regression}
\usage{
concentratedOLS3(y_dm, X_dm, ind1, ind2, ind3)
}
\arguments{
\item{y_dm}{Demeaned dependent variable}

\item{X_dm}{Demeaned explanatory variables matrix}

\item{ind1}{Indicator vector for regime 1 (integer or numeric, 1 = in regime)}

\item{ind2}{Indicator vector for regime 2}

\item{ind3}{Indicator vector for regime 3}
}
\value{
List with beta1, beta2, beta3, fitted, resid, sigma2
}
\description{
Estimates three sets of slope coefficients (beta1, beta2, beta3) by
ordinary least squares applied to each regime's sub-sample.
}
