% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeFE}
\alias{removeFE}
\title{Remove Fixed Effects via Within-Group Demeaning}
\usage{
removeFE(y, X, id)
}
\arguments{
\item{y}{Numeric vector of dependent variable}

\item{X}{Numeric matrix of explanatory variables}

\item{id}{Vector of cross-sectional identifiers}
}
\value{
List with demeaned y and X matrices
}
\description{
Remove Fixed Effects via Within-Group Demeaning
}
