% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robustSE}
\alias{robustSE}
\title{Robust Standard Errors (HC sandwich estimator)}
\usage{
robustSE(X, resid, type = "HC3")
}
\arguments{
\item{X}{Design matrix}

\item{resid}{Residual vector}

\item{type}{One of "HC0", "HC1", "HC2", "HC3"}
}
\value{
Vector of standard errors
}
\description{
Robust Standard Errors (HC sandwich estimator)
}
