% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validatePanel}
\alias{validatePanel}
\title{Validate Panel Data Structure}
\usage{
validatePanel(data, id, time)
}
\arguments{
\item{data}{Data frame}

\item{id}{Cross-sectional ID variable name}

\item{time}{Time variable name}
}
\value{
List with panel structure information
}
\description{
Validate Panel Data Structure
}
