% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validatePanel}
\alias{validatePanel}
\title{Validate Panel Data Structure}
\usage{
validatePanel(data, id, time)
}
\arguments{
\item{data}{Data frame}

\item{id}{Cross-sectional ID variable name}

\item{time}{Time variable name}
}
\value{
List with fields \code{n_units}, \code{n_periods}, \code{n_obs},
  \code{balanced} (logical), \code{has_missing} (logical)
}
\description{
Validate Panel Data Structure
}
